Generate invoice from bank_statement_line product
- Refactor of code between file completion_rules.py and statement.py
- Generate invoices from products linked to bank statement lines
